fix: :white_check_mark: Fixes a number of bugs and passes the first test on event notification.
This first test aim to test if the notification produced by the FileWatcher to the sender are exactly what we expect to see when creating, writing, renaming, removing and moving files and folders. To pass the test a number of bugs releated to the Rename and Move/Remove operations are now solved.