feat: :sparkles: Introduces client side Server status monitoring for each registered service and overall connection state
Added ServerStatus type into the status.go utils source code which
represents the status of each registered service and the overall server
status. Implemented also a reconnection mechanism whenever the
connection goes down or the overall server status appears to be
NOT_SERVING, with a backoff strategy up to 30 seconds. Also adds a new
service in the service.proto called Services which query the server for
all registered services.